Reorganise tree and start working on Windows support
Previous way to wrap platform-independent found to be impractical
for sockets interface which is close to same between platforms so
posix/socket.cpp will be moved to internal/socket.cpp when Windows
support will be finished.

file(GLOB ...) advantages overweight drawbacks in our case (through
this change may be reverted in some future).